Revert "A_1の完成とゴールの実装"
This reverts commit a9ad7b4a378b3d3eaa22daf5e408ce0fa57e4ecf.

+++ CASE_MainProject/Assets/Project/Program/GameMain/Other/ObjectFreeRotation.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class ObjectFreeRotation : MonoBehaviour
 {
     [SerializeField]
-    private float ObjectSpeed;
+    private int ObjectSpeed;
 
     [SerializeField]
     private int RotationX;
